+++ lib/dynamic_widgets/back_drop_filter.dart
@@ -10,11 +10,9 @@ final class JsonWidget$BackDropFilter extends JsonWidget {
   @override
   Widget fromJson(Map<String, dynamic> json, {List<InAppArgument>? arguments}) => BackdropFilter(
     filter: ImageFilter.blur(
-      sigmaY: _double(json['sigmaY']) ?? 0,
-      sigmaX: _double(json['sigmaX']) ?? 0,
+      sigmaY: parseDouble(json['sigmaY']) ?? 0,
+      sigmaX: parseDouble(json['sigmaX']) ?? 0,
     ),
-    child: JsonWidget.fromType(context: context, json: json['child'], arguments: arguments),
+    child: childOrEmpty(json['child'], arguments: arguments),
   );
-
-  double? _double(dynamic v) => v == null ? null : double.tryParse(v.toString());
 }